backend/requirements.txt
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AWS Bedrock for LLM
boto3>=1.34.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Form Data Support
python-multipart>=0.0.6

# Email Validation (used by Pydantic EmailStr)
email-validator>=2.0.0

# Pydantic Settings (for BaseSettings)
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0
